requirements.txt
androguard~=4.1.3

pillow~=11.3.0